% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ais_ch03_ha}
\alias{ais_ch03_ha}
\title{Subset of Ha et al. (2010) data}
\format{A data frame with 20 rows and 13 variables:
\describe{
  \item{ID}{Participant ID}
  \item{ageYears}{Age of the participant, years}
  \item{Sex}{Biological sex of the individual}
  \item{HighSalary}{Rating (out of 10) of the importance of a high salary in a potential romantic partner}
  \item{FinishedEducation}{Rating (out of 10) of the importance of a potential romantic partner having finished their education}
  \item{Kind}{Rating (out of 10) of the importance of a potential romantic partner being kind}
  \item{Humour}{Rating (out of 10) of the importance of a potential romantic partner having a sense of humour}
  \item{Ambitious}{Rating (out of 10) of the importance of a potential romantic partner being ambitious}
  \item{WantsChildren}{Rating (out of 10) of the importance of a potential romantic partner wanting children}
  \item{Romantic}{Rating (out of 10) of the importance of a potential romantic partner being romantic}
  \item{AttractiveAppearance}{Rating (out of 10) of the importance of a potential romantic partner being physically attractive}
  \item{Creativity}{Rating (out of 10) of the importance of a potential romantic partner being creative}
  \item{Honest}{Rating (out of 10) of the importance of a potential romantic partner being honest}
}}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
ais_ch03_ha
}
\description{
A dataset containing a random sample of 20 female participants from Ha, T.,
Overbeek, G., & Engels, R. C. M. E. (2010). Effects of Attractiveness and
Social Status on Dating Desire in Heterosexual Adolescents: An Experimental
Study. Archives of Sexual Behavior, 39(5), 1063-1071. doi: 10.1007/s10508-009-9561-z
}
\keyword{datasets}
