% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ais_ch05_jigsaw}
\alias{ais_ch05_jigsaw}
\title{JIG:SAW employee data.}
\format{A data frame with 240 rows and 7 variables:
\describe{
  \item{ID}{employee ID}
  \item{Employee}{whether or not the employee works for JIG:SAW}
  \item{Job_Type}{Categories of employee}
  \item{Footspeed}{Footspeed of the individual, miles per hour}
  \item{Strength}{Maximal push force of the individual, Newtons}
  \item{Vision}{Visual accuity, Newtons}
  \item{Sex}{Biological sex of the individual}
}}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
ais_ch05_jigsaw
}
\description{
A dataset containing information about the characteristics of different groups of employees within
the corporation called JIG:SAW and non-employees
}
\keyword{datasets}
