% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calcite_dat}
\alias{calcite_dat}
\title{Calcite mask data}
\format{A tibble with 20 rows and 3 variables:
\describe{
  \item{id}{Participant ID}
  \item{actual}{How many of the 20 people the participant correctly identified when a picture of the 'actual' person was transmitted to the participant's ID chip}
  \item{different}{How many of the 20 people the participant correctly identified when a picture of a 'different' person was transmitted to the participant's ID chip}
}}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
calcite_dat
}
\description{
A dataset relating to the following fictitious experiment, which is an extension of the invisibility paste data described elsewhere:
"... we asked whether we could manipulate the face perceived by the participant when viewing someone wearing calcite invisibility paste.
Participants were all so-called Chippers, who had ID chips installed in their brains. Each participant met 40 people who had calcite paste
on their face. For each of these people we took a photo of their face (the 'actual' face), and found a photo of a different face matched for
sex and age (the 'different' face). We could send these images to the participant's visual cortex using their ID chip.
For 20 of the encounters an image of the person's 'actual' face was transmitted to the participant's ID chip during the encounter, for the remaining
20 encounters the 'different' face was sent to the chip. During each encounter the participant was shown the photos of the actual and different face
and pointed to the one depicting whom they thought they had just met. As in Experiment 1, participants scored a point for each person they correctly
identified: a score of 0 would mean that they never chose the correct photo, a score of 20 would mean that they always chose the correct photo."
The data are taken from Chapter 15 of Field, A. P. (2016). An adventure in statistics: the reality enigma. London: Sage.
}
\keyword{datasets}
