% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{garlic_dat}
\alias{garlic_dat}
\title{Zombie garlic data}
\format{A tibble with 10 rows and 3 variables:
\describe{
  \item{id}{Victim ID}
  \item{latency}{The time taken for a zombie to approach the victim in seconds}
  \item{garlic}{The number of bulbs of garlic worn by the victim}
}}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
garlic_dat
}
\description{
A dataset relating to whether zombies can be repelled by garlic. It is taken from Chapter 14 of Field, A. P. (2016). An adventure in statistics: the reality enigma. London: Sage.
}
\keyword{datasets}
