% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ha_dat}
\alias{ha_dat}
\title{Subset of Ha et al. (2010) data}
\format{A data frame with 20 rows and 13 variables:
\describe{
  \item{id}{Participant ID}
  \item{age}{Age of the participant, years}
  \item{Sex}{Biological sex of the individual}
  \item{hi_salary}{Rating (out of 10) of the importance of a high salary in a potential romantic partner}
  \item{fin_ed}{Rating (out of 10) of the importance of a potential romantic partner having finished their education}
  \item{kind}{Rating (out of 10) of the importance of a potential romantic partner being kind}
  \item{humour}{Rating (out of 10) of the importance of a potential romantic partner having a sense of humour}
  \item{ambitious}{Rating (out of 10) of the importance of a potential romantic partner being ambitious}
  \item{wants_child}{Rating (out of 10) of the importance of a potential romantic partner wanting children}
  \item{romantic}{Rating (out of 10) of the importance of a potential romantic partner being romantic}
  \item{attractive}{Rating (out of 10) of the importance of a potential romantic partner being physically attractive}
  \item{creativity}{Rating (out of 10) of the importance of a potential romantic partner being creative}
  \item{honest}{Rating (out of 10) of the importance of a potential romantic partner being honest}
}}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
ha_dat
}
\description{
A dataset containing a random sample of 20 female participants from Ha, T.,
Overbeek, G., & Engels, R. C. M. E. (2010). Effects of Attractiveness and
Social Status on Dating Desire in Heterosexual Adolescents: An Experimental
Study. Archives of Sexual Behavior, 39(5), 1063-1071. doi: 10.1007/s10508-009-9561-z
}
\keyword{datasets}
