% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{implant_dat}
\alias{implant_dat}
\title{Memory implanting data}
\format{A tibble with 5 rows and 4 variables:
\describe{
  \item{id}{Participant ID}
  \item{no_interference}{How many of the 10 unusual events were remembered when there was no interference}
  \item{verbal}{How many of the 10 unusual events were remembered when verbal information was used as interference}
  \item{visual}{How many of the 10 unusual events were remembered when visual images were used as interference}
}}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
implant_dat
}
\description{
A dataset relating to the following fictitious experiment:
"We wanted to see whether our failure to replace memories was due to the type of conflicting information used. Five participants engaged in
three different scripted encounters with strangers during which 10 unusual things happened. After each encounter participants were asked to
privately recall what had happened. During the period of recall for one encounter they were left alone (control), during the recall of another
their ID chip was used to implant a verbal description of the encounter in which the 10 unusual events were replaced with different events, and
during the recall of the final encounter their ID chip was used to implant visual images of 10 different things to what they actually experienced.
Each participant, therefore, experienced three encounters and a different type of recall manipulation followed each one; the type of manipulation
associated with a particular encounter was counterbalanced. Finally, participants recalled the event for a second time and we measured how many
of the 10 unusual events that they originally experienced that they could recall."
The data are taken from Chapter 16 of Field, A. P. (2016). An adventure in statistics: the reality enigma. London: Sage.
}
\keyword{datasets}
