% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mask_dat}
\alias{mask_dat}
\title{Invisibility paste data}
\format{A tibble with 28 rows and 3 variables:
\describe{
  \item{id}{Participant ID}
  \item{mask}{Whether the person was assigned to the calcite mask condition or the invisibility paste condition}
  \item{recognition}{How many of the 20 people they correctly identified}
}}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
mask_dat
}
\description{
A dataset relating to the following fictitious experiment:
"Twenty-eight participants were tested individually. Each participant sat on a revolving chair in the middle of a circle of 20 people.
The 20 people all wore a light deflecting aid on their face. For half of the participants (N = 14) the 20 people wore a standard calcite-based mask,
whereas the other half (N = 14) saw people wearing ‘invisibility paste’, a transparent cream containing calcite micro-crystals.
Participants looked at each person and were given two photographs, one depicting the person they were looking at and one depicting a different
person matched for age and sex. Participants had to decide which person they were looking at, before turning to the next person and being given
two different photographs. Participants scored a point for each person they identified correctly: a score of 0 would mean they always chose the
incorrect photo, a score of 20 would mean that they always chose the correct photo."
The data are taken from Chapter 15 of Field, A. P. (2016). An adventure in statistics: the reality enigma. London: Sage.
}
\keyword{datasets}
