% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_advent.R
\name{read_advent}
\alias{read_advent}
\title{read_advent}
\usage{
read_advent(day = 1, cookie = NULL, year = 2021)
}
\arguments{
\item{day}{Numeric indicating which day needs to be read}

\item{cookie}{Not required if coookie set as an environment variable. A session cookie retrieved as shown in the readme.}

\item{year}{Year of the problem. By default 2021}
}
\description{
Read problem data from the Advent of Code
}
\examples{
\dontrun{
# Feed your cookie as environment variable
Sys.setenv("ADVENT_COOKIE" = "<paste_your_cookie>")
read_advent(day = 1)

# Feed the cookie as a parameter
cookie <- "<paste_your_cookie>"
read_advent(day = 1, cookie = cookie)
}
}
