% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_advent_of_code.R
\name{read_advent_of_code}
\alias{read_advent_of_code}
\title{read_advent_of_code}
\usage{
read_advent_of_code(day = 1, cookie = NULL)
}
\arguments{
\item{day}{Numeric indicating which day needs to be read}

\item{cookie}{Not required if coookie set as an environment variable. A session cookie retrieved as shown in the readme.}
}
\description{
Read problem data from the Advent of Code
}
\examples{
\dontrun{
# Feed your cookie as environment variable
Sys.setenv("ADVENT_COOKIE" = "<paste_your_cookie>")
read_advent_of_code(day = 1)

# Feed the cookie as a parameter
cookie <- "<paste_your_cookie>"
read_advent_of_code(day = 1, cookie = cookie)
}
}
