% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taser_dat}
\alias{taser_dat}
\title{Zombie immobility data}
\format{
A tibble with 156 rows and 4 variables:
\describe{
  \item{id}{Victim ID}
  \item{immobility}{The duration of the zombie's immobility, seconds}
  \item{r_tms}{The frequency of rTMS, Hz}
  \item{taser}{The output of the taser, kV}
}
}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
taser_dat
}
\description{
A dataset relating to the degree to which zombies can be rendered immobile by repetitive transcranial magnetic stimulation (rTMS) and zapping with a taser.
It is taken from Chapter 14 of Field, A. P. (2016). An adventure in statistics: the reality enigma. London: Sage.
}
\keyword{datasets}
