% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teddy_dat}
\alias{teddy_dat}
\title{Teddy therapy data.}
\format{
A tibble with 220 rows and 4 variables:
\describe{
  \item{id}{Participant ID}
  \item{group}{whether the participant cuddled a teddy bear or the cardboard box that contained the teddy bear}
  \item{study_n}{Factor that distinguishes the two studies in the data set. The first was based on a total N of 20 and the second was based on a total N of 200}
  \item{self_esteem}{Self-reported self-esteem scores}
}
}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
teddy_dat
}
\description{
A dataset based on two ficticious studies looking at whether cuddling a teddy bear (compared to the cardboard box that the teddy was packaged in) affects self-reported self-esteem.
It is taken from Chapter 10 of Field, A. P. (2016). An adventure in statistics: the reality enigma. London: Sage.
}
\keyword{datasets}
