% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhang_female_dat}
\alias{zhang_female_dat}
\title{Subset of Zhang data (female sample N = 20)}
\format{A tibble with 20 rows and 2 variables:
\describe{
  \item{name}{Specifies whether participants completed the test under their own name or a fake name}
  \item{accuracy}{Accuracy on a maths test, percent}
}}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
zhang_female_dat
}
\description{
A dataset extracted from Ha, T., Overbeek, G., & Engels, R. C. M. E. (2010). Effects of Attractiveness and Social Status on Dating Desire in Heterosexual Adolescents: An Experimental Study. Archives of Sexual Behavior, 39(5), 1063-1071. doi:10.1007/s10508-009-9561-z.
The study looked at accuracy on a maths test when performed under the participant's own name or a fake name. It is taken from Chapter 15 of Field, A. P. (2016). An adventure in statistics: the reality enigma. London: Sage.
}
\keyword{datasets}
