% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jig_dat}
\alias{jig_dat}
\title{JIG:SAW employee data.}
\format{A data frame with 240 rows and 7 variables:
\describe{
  \item{id}{employee ID}
  \item{employee}{whether or not the employee works for JIG:SAW}
  \item{job_type}{Categories of employee}
  \item{footspeed}{Footspeed of the individual, miles per hour}
  \item{strength}{Maximal push force of the individual, Newtons}
  \item{vision}{Visual accuity, Newtons}
  \item{sex}{Biological sex of the individual}
}}
\source{
\url{https://www.discoveringstatistics.com/books/an-adventure-in-statistics/}
}
\usage{
jig_dat
}
\description{
A dataset containing information about the characteristics of different groups of employees within
the corporation called JIG:SAW and non-employees
}
\keyword{datasets}
